+++ lib/cli/index.js
@@ -13,7 +13,7 @@ import {
 
 import { loadConfig } from '../core/config.js';
 import * as compile from '../core/compile/index.js';
-import { logger } from '../core/logger.js';
+import logger from '../core/logger.js';
 import { debugStats } from '../core/utils.js';
 
 import { dev } from './dev.js';
