+++ lib/cli/watch.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import chokidar from 'chokidar';
 import chalk from 'chalk';
-import { logger } from '../core/logger.js';
+import logger from '../core/logger.js';
 import * as compile from '../core/compile/index.js';
 import { findDependentPages } from '../core/dependencies.js';
 import { writePageFiles, writeEntryFile } from '../core/files.js';
@@ -81,23 +81,14 @@ export function watch(_cayo) {
 }
 
 async function handleTemplate(_cayo) {
-  try {
-    await compile.template(_cayo);
-    for (const [key] of _cayo.pages) {
-      await handlePage(key, _cayo);
-    }
-  } catch (err) {
-    logger.error(err);
+  await compile.template(_cayo);
+  for (const [key] of _cayo.pages) {
+    await handlePage(key, _cayo);
   }
 }
 
 async function handleCayo(filepath, _cayo) {
-  try {
-    // await compile.cayos(null, _cayo);
-    await handleComponent(filepath, _cayo);
-  } catch (err) {
-    logger.error(err);
-  }
+  await handleComponent(filepath, _cayo);
 }
 
 async function handleComponent(filepath, _cayo) {
@@ -121,14 +112,7 @@ async function handlePage(filepath, _cayo) {
   const page = (await compile.pages([filepath], _cayo))[0];
   pages.set(filepath, page);
   await page.render(_cayo, { load: true });
-  
-  // Recompile any cayos rendered by this page
-  // const cayos = [..._cayo.stats.cayoComponents].filter(cayo => {
-  //   if (cayo.pages.has(filepath)) return cayo;
-  // });
-  // await compile.cayos(cayos, _cayo);
 
-  // for debugging
   if (_cayo.config.debug) {
     debugStats(_cayo);
   }
