+++ lib/components/cayo-warnings.js
@@ -30,7 +30,7 @@ function checkBadProps(warnings, src, keys) {
     warnings.badProps = {
       title: `Bad props`,
       src,
-      message: `Unserializable props found: ${propsStr}. Cayo component props must be serializable in order to be passed during hydration.`,
+      message: `Unserializable props found: ${propsStr}. Cayo component props must be serializable in order to be used for hydration.`,
       log: `Unserializable instance props found: ${propsStr}.`,
     }
   }
