+++ lib/core/codegen.js
@@ -1,4 +1,6 @@
 import path from 'path';
+import chalk from 'chalk';
+import logger from './logger.js';
 
 // Generate runtime helper that finds and parses a component instance's prop data
 export function generateGetProps() {
@@ -86,4 +88,75 @@ export function generateComponentInstance(cayoId, componentName) {
 // Generate the code to import the the generated Cayo Runtime file
 export function generateCayoRuntimeImport() {
   return `import { default as renderCayos } from './cayo-runtime.js';\n`
+}
+
+// TODO: add links to relevant docs
+// Generate console warnings that need to show in the browser
+export function generateRuntimeIssuesScript(runtime, issues, type) {
+  const { config, document, page } = runtime;
+  const script = document.createElement('script');
+
+  // Formattign and label stuff
+  let prefix = 'Warning';
+  let f_log = (str) => str;
+  let f_prefix = prefix
+  let consoleType = 'log';
+  if (type === 'warning') {
+    prefix = 'Warning';
+    f_log = chalk.yellow;
+    f_prefix = chalk.yellow.bold(prefix);
+    consoleType = 'warn';
+  } else if (type === 'error') {
+    prefix = 'Error';
+    f_log = chalk.redBright;
+    f_prefix = chalk.redBright.bold(prefix);
+    consoleType = 'error';
+  }
+  
+  const { 
+    cayos: cayoIssues, 
+    page: pageIssues,
+  } = issues;
+
+  script.innerHTML = `/* ${prefix}s for this page. See issues in console. */\n`;
+
+  // Handle issues derived during initial compilation of the cayo instance
+  if (Object.keys(cayoIssues).length > 0) {
+    for (const cayoId in cayoIssues) {
+      let message = `Cayo ${prefix}: Cayo '${cayoId}' has runtime issues.\n\n`;
+      for (const key in cayoIssues[cayoId]) {
+        const issue = cayoIssues[cayoId][key];
+        message += `${issue.title}: ${issue.message}\n\n`;
+        if (cayoId !== 'undefined') {
+          message += `Hint: review instances of <Cayo src="${issue.src}"> intended to be rendered on page '${page.sourcePath.replace(config.pages, '')}'\n`;
+        } else {
+          message += `Hint: review instances of <Cayo src={<undefined>}> or <Cayo> without a src prop intended to be rendered on page '${page.sourcePath.replace(config.pages, '')}'\n`;
+        }
+        if (issue.log) {
+          logger.log.info(
+            `${f_log(`${f_prefix}: ${issue.log}`)} ${chalk.dim(`${page.name}`)}`, 
+            { timestamp: true, clear: false, }
+          );
+        }
+      }
+      script.innerHTML += `console.${consoleType}(\`${message}\`);\n`;
+    }
+  }
+  // Handle issues derived from parsing the rendered HTML page
+  if (Object.keys(pageIssues).length > 0) {
+    for (const key in pageIssues) {
+      const issue = pageIssues[key];
+      let message = `${prefix}: page '${page.name}' has runtime issues.\n\n`;
+      message += `${issue.title}: ${issue.message}\n\n`;
+      script.innerHTML += `console.${consoleType}(\`${message}\`);\n`;
+      if (issue.log) {
+        logger.log.info(
+          `${f_log(`${f_prefix}: ${issue.log}`)} ${chalk.dim(`${page.name}`)}`, 
+          { timestamp: true, clear: false, }
+        );
+      }
+    }
+  }
+
+  return script;
 }
