+++ lib/core/compile/pages.js
@@ -5,7 +5,7 @@ import path from 'path';
 import { handleDependencies } from '../dependencies.js';
 import { Page } from '../page.js';
 import { build } from '../bundle.js';
-import { logger } from '../logger.js';
+import logger from '../logger.js';
 
 export async function compilePages(pageModulePaths, _cayo) {
   const compiledPages = [];
