+++ lib/core/files.js
@@ -1,7 +1,7 @@
 import fs from 'fs-extra';
 import path from 'path';
 import chalk from 'chalk';
-import { logger } from './logger.js';
+import logger from './logger.js';
 import { processEntrySource } from './entry.js';
 
 // Write file content for a page
