+++ lib/core/logger.js
@@ -17,7 +17,9 @@ const errorLogger = (err) => {
   if (err.cause) console.error(chalk.red.bold('> Cause:'), chalk.red.bold(err.cause));
 }
 
-export const logger = { 
+const logger = { 
   log: infoLogger,
   error: errorLogger,
 }
+
+export default logger;
