+++ lib/core/render/prerender.js
@@ -1,11 +1,10 @@
-import fs, { pathExists } from 'fs-extra';
+import fs from 'fs-extra';
 import path from 'path';
 import { JSDOM } from 'jsdom';
-import chalk from 'chalk';
-import { Renderer } from './renderer.js';
-import { logger } from '../logger.js';
-import { generateCayoRuntime } from '../codegen.js';
+import logger from '../logger.js';
 import { generateCayoComponentId } from '../utils.js';
+import { Renderer } from './renderer.js';
+import { generateCayoRuntime, generateRuntimeIssuesScript } from '../codegen.js';
 import { compileCayos } from '../compile/cayos.js';
 
 export async function prerender(page, _cayo) {
@@ -66,6 +65,7 @@ export async function processPage(content, page, _cayo, logger) {
         warnings.cayos[id] = JSON.parse(el.dataset.cayoWarn);
       }
       
+      // Add cayos to stats for dependency handling
       let absoluteSrc = path.resolve(config.components, src);
       if (_cayo.stats.cayoComponents[name]) {
         _cayo.stats.cayoComponents[name].src = absoluteSrc;
@@ -78,12 +78,15 @@ export async function processPage(content, page, _cayo, logger) {
       }
       _cayo.stats.dependencies.pages[page.sourcePath].add(absoluteSrc);      
 
+      // Compile cayos & keep track of them for this page
       const [cayo] = await compileCayos({ [name]: _cayo.stats.cayoComponents[name] }, _cayo);
       if (cayoInstances[name]) {
         cayoInstances[name].push(id);
       } else {
         cayoInstances[name] = [id];
       }
+      // Include the assets needed by a cayo
+      // CSS is the only thing currently output other than the component JS
       if (!cayoAssets[name]) {
         cayoAssets[name] = { 
           css: { 
@@ -134,9 +137,7 @@ export async function processPage(content, page, _cayo, logger) {
         log: `Entry file '${entryScriptSrc}' does not exist.`,
       }
       // Remove the entry file script because the file doesn't exist
-      if (entryScriptPlaceholder) {
-        entryScript.remove();
-      }
+      entryScript.remove();
     } else {
       entry.path = absoluteEntrySrcPath;
       _cayo.stats.dependencies.pages[page.sourcePath].add(absoluteEntrySrcPath);
@@ -152,9 +153,7 @@ export async function processPage(content, page, _cayo, logger) {
       }
     } else {
       // Remove the entry point script tag because the page doesn't need any JS
-      if (entryScript) {
-        entryScript.remove();
-      }
+      entryScript.remove();
     }
   }
 
@@ -162,26 +161,27 @@ export async function processPage(content, page, _cayo, logger) {
   const cayoRuntimePath = page.url === '/' ? `/index.js` : `${page.url}/index.js`;
   js = generateCayoRuntime(cayoInstances, cayoRuntimePath, _cayo);
 
+  // Indicate that we need to make this entry ready to render cayos later,
+  // if there are cayo instances on this page
   if (cayoIds.length > 0) {
-    // Indicate that we need to make this entry ready to render cayos later
     entry.renderCayos = true;
   }
   
-  // Handle runtime warnings  
+  // Append runtime warnings
   const warningScript = generateRuntimeIssuesScript({ config, document, page }, warnings, 'warning');
   if (config.mode === 'development') {
     document.body.appendChild(warningScript);
   }
-  
+  // Append runtime errors
   const errorScript = generateRuntimeIssuesScript({ config, document, page }, errors,  'error');
   if (config.mode === 'development') {
     document.body.appendChild(errorScript);;
   }
 
-  // Construct the correct HTML string based on the document tags
-  // that were rendered from the source (jsdom wraps the source HTML in a document,
+  // Construct the correct HTML string based on the document structure
+  // that was rendered from the source (jsdom wraps the source HTML in a document,
   // which always includes `html`, `head`, and `body`, even if the source doesn't)
-  const tags = findDocumentTags(content.html);
+  const tags = tagsExist(content.html);
   let processedHTML = '';
   if (tags.html) {
     processedHTML = document.documentElement.outerHTML;
@@ -210,84 +210,13 @@ export async function processPage(content, page, _cayo, logger) {
   };
 }
 
-function findDocumentTags(source) {
+function tagsExist(source) {
   const head = source.match(/\<head[\s\S]*\>(?<innerHTML>[\s\S]*)\<\/head\>/g);
   const body = source.match(/\<body[\s\S]*\>(?<innerHTML>[\s\S]*)\<\/body\>/g);
   const html = source.match(/\<html[\s\S]*\>(?<innerHTML>[\s\S]*)\<\/html\>/g);
-
   return {
     head: head === null ? false : true,
     body: body === null ? false : true,
     html: html === null ? false : true,
   };
-}
-
-// TODO: move to codegen.js
-// TODO: add links to relevant docs
-// Generate console warnings that need to show in the browser
-function generateRuntimeIssuesScript(runtime, issues, type) {
-  const { config, document, page } = runtime;
-  const script = document.createElement('script');
-
-  // Formattign and label stuff
-  let prefix = 'Warning';
-  let f_log = (str) => str;
-  let f_prefix = prefix
-  let consoleType = 'log';
-  if (type === 'warning') {
-    prefix = 'Warning';
-    f_log = chalk.yellow;
-    f_prefix = chalk.yellow.bold(prefix);
-    consoleType = 'warn';
-  } else if (type === 'error') {
-    prefix = 'Error';
-    f_log = chalk.redBright;
-    f_prefix = chalk.redBright.bold(prefix);
-    consoleType = 'error';
-  }
-  
-  const { 
-    cayos: cayoIssues, 
-    page: pageIssues,
-  } = issues;
-
-  script.innerHTML = `/* ${prefix}s for this page. See issues in console. */\n`;
-
-  // Handle issues derived during initial compilation
-  if (Object.keys(cayoIssues).length > 0) {
-    for (const cayoId in cayoIssues) {
-      let message = `Cayo ${prefix}: Cayo '${cayoId}' has runtime issues.\n\n`;
-      for (const key in cayoIssues[cayoId]) {
-        const issue = cayoIssues[cayoId][key];
-        message += `${issue.title}: ${issue.message}\n\n`;
-        if (cayoId !== 'undefined') {
-          message += `Hint: review instances of <Cayo src="${issue.src}"> intended to be rendered on page '${page.sourcePath.replace(config.pages, '')}'\n`;
-        } else {
-          message += `Hint: review instances of <Cayo src={<undefined>}> or <Cayo> without a src prop intended to be rendered on page '${page.sourcePath.replace(config.pages, '')}'\n`;
-        }
-        if (issue.log) {
-          logger.log.info(
-            `${f_log(`${f_prefix}: ${issue.log}`)} ${chalk.dim(`${page.name}`)}`, 
-            { timestamp: true, clear: false, }
-          );
-        }
-      }
-      script.innerHTML += `console.${consoleType}(\`${message}\`);\n`;
-    }
-  }
-  // Handle issues based on parsing the rendered HTML
-  if (Object.keys(pageIssues).length > 0) {
-    for (const key in pageIssues) {
-      const issue = pageIssues[key];
-      let message = `${prefix}: page '${page.name}' has runtime issues.\n\n`;
-      message += `${issue.title}: ${issue.message}\n\n`;
-      script.innerHTML += `console.${consoleType}(\`${message}\`);\n`;
-      logger.log.info(
-        `${f_log(`${f_prefix}: ${issue.log}`)} ${chalk.dim(`${page.name}`)}`, 
-        { timestamp: true, clear: false, }
-      );
-    }
-  }
-
-  return script;
 }
